Inserito il contesto di Prom tra le varie conversioni, aggiunta la map dei marking di ogni transazioni, velocizzata il metodo isCutOff, spostamento dei file LocalConfiguration in package a parte, inseriti commenti e cambiati i nomi di alcune variabili per migliorare la leggibilita
