+++ src/org/processmining/plugins/unfolding/BPMN2Unfolding_Plugin.java
@@ -5,7 +5,7 @@ import org.processmining.framework.plugin.PluginContext;
 import org.processmining.framework.plugin.annotations.Plugin;
 import org.processmining.models.graphbased.directed.bpmn.BPMNDiagram;
 import org.processmining.models.graphbased.directed.petrinet.Petrinet;
-import org.processmining.plugins.converters.bpmn2pn.BPMN2PetriNetConverter;
+import org.processmining.plugins.converters.bpmn2pn.BPMN2WSConverter;
 import org.processmining.support.unfolding.IdentificationMap;
 
 /**
@@ -15,40 +15,67 @@ import org.processmining.support.unfolding.IdentificationMap;
  */
 public class BPMN2Unfolding_Plugin {
 	@Plugin(
-		name = "Convert BPMN diagram to Unfolding net", 
-		parameterLabels = {"Petri net"},
-		returnLabels = {  "Identification Map","Petri net" }, 
-		returnTypes = {  IdentificationMap.class,Petrinet.class }, 
+		name = "BCS BPMN to Unfolding", 
+		parameterLabels = {"BPMNDiagram"},
+		returnLabels = {"Identification Map","Petri net"}, 
+		returnTypes = {IdentificationMap.class, Petrinet.class}, 
 		userAccessible = true, 
 		help = "Convert BPMN diagram to Unfolding net"
 		)
 	@UITopiaVariant(
-		affiliation = "Università di Pisa", 
+		affiliation = "University of Pisa", 
 		author = "Daniele Cicciarella", 
 		email = "cicciarellad@gmail.com"
 		)
 	public Object[] convert(PluginContext context, BPMNDiagram bpmn) 
 	{
 		Petrinet petrinet;
+		BPMN2WSConverter bpmn2Petrinet;
+		PetriNet2Unfolding petrinet2Unfolding;
 		Object[] unfolding;
-		context.getProgress().setMinimum(0);
-		context.getProgress().setMaximum(3);
 		
-		/* Converte BPMNDiagram in Petri net */
-		BPMN2PetriNetConverter bpmn2petrinet = new BPMN2PetriNetConverter(bpmn);
-		bpmn2petrinet.convert();
-		petrinet = bpmn2petrinet.getPetriNet();
-		context.log("BPMN convertito in rete di Petri");
-		context.getProgress().inc();
+		/* Settiamo la barra progressiva */
+		setProgress(context, 0, 3);
+		
+		/* Converte il BPMN in rete di Petri */
+		writeLog(context, "Conversion of the BPMN in Petri net...");
+		bpmn2Petrinet = new BPMN2WSConverter(bpmn);
+		bpmn2Petrinet.convert();
+		petrinet = bpmn2Petrinet.getPetriNet();
 
-		/* Converte Petri net in una Occurrence net con Unfolding */
-		PetriNet2Unfolding petrinet2unfolding = new PetriNet2Unfolding(context, petrinet);
-		unfolding = petrinet2unfolding.convert();
-		context.getProgress().inc();
+		/* Converte la rete di Petri nella rete di unfolding */
+		writeLog(context, "Conversion of the Petri net in Unfolding net...");
+		petrinet2Unfolding = new PetriNet2Unfolding(context, petrinet);
+		unfolding = petrinet2Unfolding.convert();
 		
 		/* Aggiungo connessione per la visualizzazione delle reti e statistiche delle rete unfoldata */
 		context.addConnection(new UnfoldingConnection((IdentificationMap)unfolding[1], petrinet,(Petrinet) unfolding[0]));
 		
 		return new Object [] {unfolding[1], unfolding[0]};
+	}
+	
+	/**
+	 * Setta gli step della barra progressiva
+	 * 
+	 * @param context contesto di ProM
+	 * @param minimun minimo valore
+	 * @param maximum massimo valore
+	 */
+	private void setProgress(PluginContext context, int minimun, int maximum)
+	{
+		context.getProgress().setMinimum(minimun);
+		context.getProgress().setMaximum(maximum);
+	}
+	
+	/**
+	 * Scrive un messaggio di log e incrementa la barra progressiva
+	 * 
+	 * @param context contesto di ProM
+	 * @param log messaggio di log
+	 */
+	private void writeLog(PluginContext context, String log)
+	{
+		context.log(log);
+		context.getProgress().inc();
 	}
 }
